migrate to neotraits: rest of datatypes that are involved in simulations
git-svn-id: https://repo.thevirtualbrain.org/svn/tvb/trunk@8977 1c0e02f0-7929-43c0-8fb3-3293bf43b0d1

+++ scientific_library/tvb/datatypes/volumes.py
@@ -39,30 +39,17 @@ methods that are associated with the volume datatypes.
 from tvb.basic.logger.builder import get_logger
 from tvb.basic.traits.neotraits import HasTraits, Attr, NArray
 
+
 LOG = get_logger(__name__)
 
 
 class Volume(HasTraits):
     """
     Data defined on a regular grid in three dimensions.
-
     """
-    origin = NArray(
-        dtype=float,
-        label="Volume origin coordinates"
-    )
-
-    voxel_size = NArray(
-        dtype=float,
-        label="Voxel size"
-    )
-    # need a triplet, xyz
-
-    voxel_unit = Attr(
-        str,
-        label="Voxel Measure Unit",
-        default="mm"
-    )
+    origin = NArray(label="Volume origin coordinates")
+    voxel_size = NArray(label="Voxel size") # need a triplet, xyz
+    voxel_unit = Attr(str, label="Voxel Measure Unit", default="mm")
 
     def _find_summary_info(self):
         summary = {"Volume type": self.__class__.__name__,
